OTH: Added a lot of comments and refactored folder structure
