+++ ProcessorOrchestrator/src/handlers/processor_socket.py
@@ -12,10 +12,10 @@ import json
 
 from tornado.websocket import WebSocketHandler
 
-from src.object_manager import objects
-from src.connections import processors
-import src.client_socket as client_socket
-import src.logger as logger
+from src.objects.object_management import objects
+from src.objects.connections import processors
+import src.handlers.client_socket as client_socket
+import src.utility.logger as logger
 
 
 class ProcessorSocket(WebSocketHandler):
