+++ ProcessorOrchestrator/src/utility/http_server.py
@@ -12,7 +12,7 @@ import ssl
 
 from tornado.httpserver import HTTPServer
 
-import src.logger as logger
+import src.utility.logger as logger
 
 
 def create_http_servers(app):
@@ -27,11 +27,11 @@ def create_http_servers(app):
     # Get ssl ready, if provided in the environment variables.
     cert = os.environ.get('SSL_CERT')
     key = os.environ.get('SSL_KEY')
-    use_tls = cert is not None and key is not None
 
     https_server = None
 
-    if use_tls:
+    # A second https server should only be created when certificates are provided.
+    if cert is not None and key is not None:
         # Create a ssl context.
         ssl_ctx = ssl.create_default_context(ssl.Purpose.CLIENT_AUTH)
         ssl_ctx.load_cert_chain(cert, key)
